Revert "[Refactor][Philippe] Modularized Modals and API Calls of Settings and Edit Account."
