+++ client/src/screens/settings/editaccount.tsx
@@ -1,10 +1,10 @@
-import React, { useState, useEffect } from 'react';
-import { StyleSheet, View, Text, TouchableOpacity, ScrollView } from 'react-native';
-import { NavioScreen } from 'rn-navio';
-import { supabase } from '@app/lib/supabase';
-import { Icon } from '@app/components/icon';
-import { EditModal } from '@app/components/editModal';
-import { fetchUserData, updateUserData } from '@app/services/api/userdetails';
+import React, {useState, useEffect} from 'react';
+import {View, Text, TextInput, Button, StyleSheet, ScrollView, TouchableOpacity, Modal} from 'react-native';
+import {DateTimePicker} from 'react-native-ui-lib';
+import {NavioScreen} from 'rn-navio';
+import {supabase} from '@app/lib/supabase';
+import {Icon, IconName} from '@app/components/icon';
+import {PickerFixed} from '@app/components/picker-fixed';
 
 export const EditAccount: NavioScreen = () => {
   const [email, setEmail] = useState('');
@@ -21,36 +21,43 @@ export const EditAccount: NavioScreen = () => {
   const [currentValue, setCurrentValue] = useState('');
 
   useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const data = await fetchUserData();
+    const fetchUserData = async () => {
+      const { data: { user }, error: authError } = await supabase.auth.getUser();
+      if (authError || !user) throw new Error("User not authenticated");
+      const { data, error } = await supabase
+        .from('UserDetails')
+        .select('email, first_name, last_name, birthdate, gender, mobile_number')
+        .eq('user_id', user.id)  
+        .limit(1)
+        .single();
+
+      if (error) {
+        console.error('Error fetching user data:', error);
+      } else if (!data) {
+        console.error('No user data found');
+      } else {
         setEmail(data.email);
         setFirstName(data.first_name);
         setLastName(data.last_name);
         setDateOfBirth(data.birthdate);
         setGender(data.gender);
         setPhoneNumber(data.mobile_number);
-      } catch (error) {
-        if (error instanceof Error) {
-          console.error(error.message);
-        } else {
-          console.error('An unknown error occurred:', error);
-        }
-      } finally {
-        setLoading(false);
       }
+      setLoading(false);
     };
 
-    fetchData();
+    fetchUserData();
   }, []);
 
-  const openModal = (field: React.SetStateAction<string>, value: React.SetStateAction<string>) => {
+  const openModal = (field, value) => {
+    console.log('Opening modal for field:', field, 'with value:', value);
     setCurrentField(field);
     setCurrentValue(value);
     setModalVisible(true);
   };
 
   const handleModalSave = async () => {
+    console.log('Saving modal data for field:', currentField, 'with value:', currentValue);
     const { data: { user } } = await supabase.auth.getUser();
     if (user) {
       let updateData = {};
@@ -58,7 +65,10 @@ export const EditAccount: NavioScreen = () => {
         case 'Email':
           setEmail(currentValue);
           updateData = { email: currentValue };
-          const { error: authError } = await supabase.auth.updateUser({ email: currentValue });
+          // Update email in authentication table
+          const { error: authError } = await supabase.auth.updateUser({
+            email: currentValue,
+          });
           if (authError) {
             console.error('Error updating email in authentication table:', authError);
             return;
@@ -66,6 +76,7 @@ export const EditAccount: NavioScreen = () => {
           break;
         case 'Password':
           setPassword(currentValue);
+          // Assuming password is stored in a different table or handled differently
           break;
         case 'First Name':
           setFirstName(currentValue);
@@ -92,15 +103,15 @@ export const EditAccount: NavioScreen = () => {
       }
 
       if (Object.keys(updateData).length > 0) {
-        try {
-          await updateUserData(user.id, updateData);
+        const { error } = await supabase
+          .from('UserDetails')
+          .update(updateData)
+          .eq('user_id', user.id);
+
+        if (error) {
+          console.error('Error updating user data:', error);
+        } else {
           console.log('Field updated successfully:', updateData);
-        } catch (error) {
-          if (error instanceof Error) {
-            console.error(error.message);
-          } else {
-            console.error('An unknown error occurred:', error);
-          }
         }
       }
     }
@@ -153,14 +164,51 @@ export const EditAccount: NavioScreen = () => {
       </TouchableOpacity>
       <View style={styles.divider} />
 
-      <EditModal
-        modalVisible={modalVisible}
-        currentField={currentField}
-        currentValue={currentValue}
-        setCurrentValue={setCurrentValue}
-        handleModalSave={handleModalSave}
-        setModalVisible={setModalVisible}
-      />
+      <Modal
+        transparent={true}
+        visible={modalVisible}
+        animationType="slide"
+        onRequestClose={() => setModalVisible(false)}
+      >
+        <View style={styles.modalContainer}>
+          <View style={styles.modalContent}>
+            <Text style={styles.modalTitle}>Edit {currentField}</Text>
+            {currentField === 'Birth Date' ? (
+              <DateTimePicker
+                accent
+                fieldStyle={{backgroundColor: 'white', borderWidth: 2, borderColor: 'grey', borderRadius: 6, padding: 4}}
+                label='Date of Birth'
+                labelStyle={{fontWeight: 'bold'}}
+                placeholder='DD/MM/YYYY'
+                placeholderTextColor={'grey'}
+                value={new Date(currentValue)}
+                onChange={(date) => setCurrentValue(date.toISOString().split('T')[0])}
+                mode="date"
+                maximumDate={new Date()}
+              />
+            ) : currentField === 'Gender' ? (
+              <PickerFixed
+                value={currentValue}
+                placeholder='Gender'
+                onValueChange={setCurrentValue}
+                items={['Male', 'Female', 'Other', 'Prefer not to say']}
+              />
+            ) : (
+              <TextInput
+                style={styles.input}
+                value={currentValue}
+                onChangeText={setCurrentValue}
+                placeholder={`Enter your ${currentField}`}
+                keyboardType={currentField === 'Phone Number' ? 'phone-pad' : 'default'}
+              />
+            )}
+            <View style={styles.modalButtonContainer}>
+              <Button title="Cancel" onPress={() => setModalVisible(false)} />
+              <Button title="Save" onPress={handleModalSave} />
+            </View>
+          </View>
+        </View>
+      </Modal>
     </ScrollView>
   );
 };
@@ -169,7 +217,7 @@ EditAccount.options = {
   title: 'Edit Account',
 };
 
-export const styles = StyleSheet.create({
+const styles = StyleSheet.create({
   container: {
     flexGrow: 1,
     padding: 16,
@@ -188,5 +236,37 @@ export const styles = StyleSheet.create({
     height: 1,
     backgroundColor: 'gray',
     marginVertical: 8,
+  },
+  input: {
+    height: 40,
+    width: 180,
+    borderColor: 'gray',
+    borderWidth: 1,
+    marginBottom: 16,
+    paddingHorizontal: 8,
+  },
+  modalContainer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: 'rgba(0, 0, 0, 0.5)',
+  },
+  modalContent: {
+    width: 300,
+    padding: 20,
+    backgroundColor: 'white',
+    borderRadius: 10,
+    alignItems: 'center',
+  },
+  modalTitle: {
+    fontSize: 18,
+    fontWeight: 'bold',
+    marginBottom: 10,
+  },
+  modalButtonContainer: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    width: '100%',
+    marginTop: 10,
   },
 });
